net/memnet: rename from net/nettest
This is just #cleanup to resolve a TODO

Also add a package doc.
